+++ modules/apps/headless/headless-admin-user/headless-admin-user-impl/src/main/java/com/liferay/headless/admin/user/internal/resource/v1_0/OrganizationResourceImpl.java
@@ -20,14 +20,14 @@ import com.liferay.headless.admin.user.dto.v1_0.Organization;
 import com.liferay.headless.admin.user.dto.v1_0.OrganizationContactInformation;
 import com.liferay.headless.admin.user.dto.v1_0.Service;
 import com.liferay.headless.admin.user.internal.dto.v1_0.converter.OrganizationResourceDTOConverter;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderAddressUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderCountryUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderEmailAddressUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderListTypeUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderPhoneUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderRegionUtil;
+import com.liferay.headless.admin.user.internal.dto.v1_0.util.ServiceBuilderWebsiteUtil;
 import com.liferay.headless.admin.user.internal.odata.entity.v1_0.OrganizationEntityModel;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderAddressUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderCountryUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderEmailAddressUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderListTypeUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderPhoneUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderRegionUtil;
-import com.liferay.headless.admin.user.internal.service.builder.ServiceBuilderWebsiteUtil;
 import com.liferay.headless.admin.user.resource.v1_0.OrganizationResource;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
